requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy
tqdm
matplotlib
pandas
Pillow
wandb
hydra-core
omegaconf
librosa
soundfile
